Reorganize and correct includes
