+++ src/camera/camera.cpp
@@ -3,4 +3,8 @@
 #include "camera.h"
 
 namespace CE3D
-{ }
+{
+
+
+
+}
