+++ src/model/point_model.h
@@ -3,15 +3,15 @@
 #ifndef CE3D_MODEL_POINT_MODEL_H
 #define CE3D_MODEL_POINT_MODEL_H
 
-#include "util/stdmacro.h"
+#include "util/stdinc.h"
 
-#include "model.h"
+#include "model/model.h"
 
 namespace CE3D
 {
 
 /**
- * TODO
+ * This is a model represented by just a set of points.
  */
 class PointModel : public Model
 {
