+++ src/test/console/console_test.cpp
@@ -6,7 +6,8 @@
 #define BOOST_TEST_MODULE Translation
 #include <boost/test/unit_test.hpp>
 
-#include "util/stdmacro.h"
+#include "util/stdinc.h"
+
 #include "console/console.h"
 
 namespace CE3D
