+++ src/test/transformation/CMakeLists.txt
@@ -1,4 +1,4 @@
-AddCE3DTest(TransformationTest 
+AddCE3DTest(TranslationTest
     ${CMAKE_CURRENT_LIST_DIR}/transformation_test
     ${CMAKE_CURRENT_LIST_DIR}/translation_test
     )
