+++ src/util/CMakeLists.txt
@@ -1,6 +1,4 @@
 list(APPEND SRC
-    ${CMAKE_CURRENT_LIST_DIR}/stdmacro
-    ${CMAKE_CURRENT_LIST_DIR}/stdmacro_c
     ${CMAKE_CURRENT_LIST_DIR}/boost_extensions
     ${CMAKE_CURRENT_LIST_DIR}/keyboard_state
     )
