+++ src/util/stdinc.h
@@ -1,0 +1,23 @@
+// This file is part of CE3D. License: GPL3
+
+#ifndef CE3D_UTIL_STDINC_H
+#define CE3D_UTIL_STDINC_H
+
+#include <cstdint>
+#include "CE3D_config.h"
+
+namespace CE3D
+{
+
+/**
+ * Data type for model vectors.
+ */
+using ModelDataType = float;
+/**
+ * Index type for model vectors.
+ */
+using ModelIdxType = std::uint32_t;
+
+}
+
+#endif /* CE3D_UTIL_STDINC_H */
