+++ src/util/stdinc_c.h
@@ -1,7 +1,7 @@
 // This file is part of CE3D. License: GPL3
 
-#ifndef CE3D_UTIL_STDMACRO_C_H
-#define CE3D_UTIL_STDMACRO_C_H
+#ifndef CE3D_UTIL_STDINC_C_H
+#define CE3D_UTIL_STDINC_C_H
 
 #include <stdint.h>
 #include "CE3D_config.h"
@@ -15,4 +15,4 @@ typedef uint8_t bool;
 #define TRUE 1
 #define FALSE 0
 
-#endif /* CE3D_UTIL_STDMACRO_C_H */
+#endif /* CE3D_UTIL_STDINC_C_H */
